+++ src/test/php/environments/errorhandler/DefaultErrorHandlerTest.php
@@ -7,22 +7,22 @@
  *
  * @package  stubbles
  */
-namespace stubbles\lang\errorhandler;
+namespace stubbles\environments\errorhandler;
 use function bovigo\assert\assert;
 use function bovigo\assert\predicate\isInstanceOf;
 use function stubbles\lang\extractObjectProperties;
 /**
- * Tests for stubbles\lang\errorhandler\DefaultErrorHandler.
+ * Tests for stubbles\environments\errorhandler\DefaultErrorHandler.
  *
- * @group  lang
- * @group  lang_errorhandler
+ * @group  environments
+ * @group  environments_errorhandler
  */
 class DefaultErrorHandlerTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * instance to test
      *
-     * @type  DefaultErrorHandler
+     * @type  \stubbles\environments\errorhandler\DefaultErrorHandler
      */
     private $defaultErrorHandler;
 
@@ -42,7 +42,7 @@ class DefaultErrorHandlerTest extends \PHPUnit_Framework_TestCase
         $properties = extractObjectProperties($this->defaultErrorHandler);
         assert(
                 $properties['errorHandlers'][0],
-                isInstanceOf(IllegalArgumentErrorHandler::class)
+                isInstanceOf(InvalidArgument::class)
         );
     }